compartmentalized sequelize tables
