+++ models/restaurant.js
@@ -1,32 +1,13 @@
-var bcrypt = require("bcryptjs");
-
 module.exports = function(sequelize, DataTypes) {
     var Restaurant = sequelize.define("Restaurant", {
-        id: {
+        resID: {
             type: DataTypes.STRING,
-            allowNull: false,
-            primaryKey: true
+            allowNull: false
         },
         name: {
             type: DataTypes.STRING,
             allowNull: false
         }
     });
-    var User = sequelize.define("User", {
-        username: {
-            type: DataTypes.STRING,
-            allowNull: false
-        },
-        password: {
-            type: DataTypes.STRING,
-            allowNull: false
-        }
-    });
-    User.prototype.validPassword = function(password) {
-        return bcrypt.compareSync(password, this.password);
-    };
-    User.addHook("beforeCreate", function(user) {
-        user.password = bcrypt.hashSync(user.password, bcrypt.genSaltSync(10), null);
-    });
-    return Restaurant, User;
+    return Restaurant;
 };
